requirements.txt
gradio
Pillow
python-dotenv
PyPDF2
camel-ai[all]